requirements.txt
zhipuai>=2.0.0
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sounddevice>=0.4.6
soundfile>=0.12.0
pydub>=0.25.1
streamlit-webrtc>=0.44.0